[RemoveRadium] Consolidate Carousel
