+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+
+- Feature: Consolidate both `Carousel` components and handle legacy.
 - Fix: No SSR for `<Modal />`.
 
 ## [2.70.0] - 2020-05-15
